+++ web/web/settings.py
@@ -79,6 +79,7 @@ OPT_ZER0M0N = False
 
 COMMENTS = web_cfg.comments.enabled
 ADMIN = web_cfg.admin.enabled
+ANON_VIEW = web_cfg.general.anon_viewable
 
 # If false run next command
 # python3 manage.py runserver 0.0.0.0:8000 --insecure
@@ -475,6 +476,8 @@ CSP_REPORT_PERCENTAGE = 0.6
 CSP_FONT_SRC = ["https://fonts.googleapis.com"]
 CSP_STYLE_SRC = ["'self'"]
 CSP_IMG_SRC = ["'self'"]
+
+RATELIMIT_ERROR_MSG = "Too many request without auth! You have exceed your free request per minute for anon users. We are researcher friendly and provide api, but if you buy a good whiskey to @doomedraven, we will be even more friendlier ;). Limits can be changed in conf/api.conf"
 
 # Hack to import local settings.
 try: